Move the Ristretto vectors to a vector file.
